Refactor Hugging Face components by moving ModelCard, ModelCardActions, and ModelCardContent in own folder. Updated import paths and consolidated model action logic into separate components.
