+++ web/src/components/hugging_face/RequiredModelsDialog.tsx
@@ -13,7 +13,7 @@ import {
 } from "@mui/material";
 import CloseIcon from "@mui/icons-material/Close";
 import { TOOLTIP_ENTER_DELAY } from "../../config/constants";
-import ModelCard from "./ModelCard";
+import ModelCard from "./model_card/ModelCard";
 import { useModelDownloadStore } from "../../stores/ModelDownloadStore";
 import { DownloadProgress } from "./DownloadProgress";
